refactor: routes to page view flows for easier navigation and better feature flow structures
