+++ android/app/google-services.json
@@ -76,6 +76,42 @@
           ]
         }
       }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:576582491170:android:bf30e8c7e0dc2b4b0ffe7b",
+        "android_client_info": {
+          "package_name": "dev.kumuly.kumuly_pocket"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "576582491170-m6rb9v4mfbkamsgm86mv22cjh0bc84tt.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyDrEnvJYsgjcAfC6C9a3CRdK5U9kKOj0Cc"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "576582491170-m6rb9v4mfbkamsgm86mv22cjh0bc84tt.apps.googleusercontent.com",
+              "client_type": 3
+            },
+            {
+              "client_id": "576582491170-2vl18784lm0lm7r84c484mjqdsnbtdf0.apps.googleusercontent.com",
+              "client_type": 2,
+              "ios_info": {
+                "bundle_id": "com.example.kumulyStore"
+              }
+            }
+          ]
+        }
+      }
     }
   ],
   "configuration_version": "1"
