+++ lib/l10n/app_en.arb
@@ -94,5 +94,6 @@
   "paymentCompleted": "Payment completed",
   "promoCodeReceived": "Promo code received",
   "checkOutGuidelinesToRedeemPromo": "Check out the guidelines on\nhow to redeem the promo.",
-  "redemptionInstructions": "Redemption instructions"
+  "redemptionInstructions": "Redemption instructions",
+  "promoCode": "Promo code"
 }
