+++ lib/router/app_router.dart
@@ -1,12 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:kumuly_pocket/features/landing/landing_screen.dart';
+import 'package:kumuly_pocket/features/promo_flow/code/promo_code_screen.dart';
+import 'package:kumuly_pocket/features/promo_flow/promo_flow.dart';
+import 'package:kumuly_pocket/features/promos/promos_screen.dart';
+import 'package:kumuly_pocket/features/receive_sats_flow/receive_sats_flow.dart';
 import 'package:kumuly_pocket/features/root/root_screen.dart';
 import 'package:kumuly_pocket/router/merchant_mode_route.dart';
 import 'package:kumuly_pocket/router/pocket_mode_route.dart';
-import 'package:kumuly_pocket/router/promos_route.dart';
-import 'package:kumuly_pocket/router/receive_sats_route.dart';
 import 'package:kumuly_pocket/router/sign_in_route.dart';
 import 'package:kumuly_pocket/router/sign_up_route.dart';
+import 'package:kumuly_pocket/view_models/promo.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:go_router/go_router.dart';
 
@@ -36,7 +39,7 @@ GoRouter appRouter(AppRouterRef ref) {
       signInRoute,
       GoRoute(
         path: '/import-account',
-        name: 'import-account',
+        name: 'import-account-flow',
         builder: (context, state) => Scaffold(
           appBar: AppBar(
             title: const Text('Import Account'),
@@ -48,9 +51,33 @@ GoRouter appRouter(AppRouterRef ref) {
         ),
       ),
       pocketModeRoute,
-      receiveSatsRoute,
-      promosRoute,
-      ...promoRoute,
+      GoRoute(
+        path: '/receive-sats',
+        name: 'receive-sats-flow',
+        builder: (context, state) => const ReceiveSatsFlow(),
+      ),
+      GoRoute(
+        parentNavigatorKey: _rootNavigatorKey,
+        path: '/promos',
+        name: 'promos',
+        builder: (context, state) => const PromosScreen(),
+      ),
+      GoRoute(
+        parentNavigatorKey: _rootNavigatorKey,
+        path: '/promo',
+        name: 'promo-flow',
+        builder: (context, state) => PromoFlow(
+          promo: state.extra as Promo,
+        ),
+      ),
+      GoRoute(
+        parentNavigatorKey: _rootNavigatorKey,
+        path: '/promo-code',
+        name: 'promo-code',
+        builder: (context, state) => PromoCodeScreen(
+          promo: state.extra as Promo,
+        ),
+      ),
       merchantModeRoute,
     ],
   );
