+++ lib/router/sign_in_route.dart
@@ -2,6 +2,7 @@ import 'package:go_router/go_router.dart';
 import 'package:kumuly_pocket/features/sign_in/insert_pin_screen.dart';
 import 'package:kumuly_pocket/features/sign_in/sign_in_screen.dart';
 
+// Todo: only show insert pin screen and make the app single-account
 final signInRoute = GoRoute(
   path: '/sign-in',
   name: 'sign-in',
